scripts: sync a couple of scripts

+++ Utilities/Scripts/slice2html.sh
@@ -18,20 +18,6 @@ function Usage {
 }
 
 #---------------------------------------------
-#                   trim
-#---------------------------------------------
-
-trim()
-{
-  local var="$*"
-# remove leading whitespace characters
-  var="${var#"${var%%[![:space:]]*}"}"
-# remove trailing whitespace characters
-  var="${var%"${var##*[![:space:]]}"}"
-  printf '%s' "$var"
-}
-
-#---------------------------------------------
 #                   is_smokeview_installed
 #---------------------------------------------
 
@@ -49,21 +35,12 @@ is_smokeview_installed()
 }
 
 #---------------------------------------------
-#                   restore_localstate
+#                   OUTPUT_SLICES
 #---------------------------------------------
 
-restore_localstate()
+OUTPUT_SLICES ()
 {
-  index=$1
-  LOCALCONFIG=$CONFIGDIR/slice2html_${input}_${index}
-  if [ -e $LOCALCONFIG ]; then
-    source $LOCALCONFIG
-    var="SLICE2HTML_${input}_${index}_min"
-    valmin=${!var}
-    var="SLICE2HTML_${input}_${index}_max"
-    valmax=${!var}
-    have_bounds=1
-  fi
+  cat $slicefilemenu | awk -F"," '{ print $1" ",$2," ",$3," ",$4}'
 }
 
 #---------------------------------------------
@@ -74,23 +51,10 @@ restore_state()
 {
   if [ -e $GLOBALCONFIG ]; then
     source $GLOBALCONFIG
-    HTMLDIR=${SLICE2HTML_HTMLDIR}
-    TIMEFRAME=${SLICE2HTML_TIMEFRAME}
-    EMAIL=${SLICE2HTML_EMAIL}
+    HTMLDIR=$SLICE2HTML_HTMLDIR
+    TIMEFRAME=$SLICE2HTML_TIMEFRAME
+    EMAIL=$SLICE2HTML_EMAIL
   fi
-}
-
-#---------------------------------------------
-#                   save_localstate
-#---------------------------------------------
-
-save_localstate()
-{
-  index=$1
-  LOCALCONFIG=$CONFIGDIR/slice2html_${input}_${index}
-  echo "#/bin/bash"                                 > $LOCALCONFIG
-  echo "SLICE2HTML_${input}_${index}_min=$valmin"  >> $LOCALCONFIG
-  echo "SLICE2HTML_${input}_${index}_max=$valmax"  >> $LOCALCONFIG
 }
 
 #---------------------------------------------
@@ -99,55 +63,12 @@ save_localstate()
 
 save_state()
 {
-  echo "#/bin/bash"                      >  $GLOBALCONFIG
-  echo "SLICE2HTML_HTMLDIR=$HTMLDIRDIR"  >> $GLOBALCONFIG
-  echo "SLICE2HTML_TIMEFRAME=$TIMEFRAME" >> $GLOBALCONFIG
-  echo "SLICE2HTML_EMAIL=$EMAIL"         >> $GLOBALCONFIG
-}
-
-#---------------------------------------------
-#                   writeini
-#---------------------------------------------
-
-
-writeini ()
-{
-ini_filename=$1
-echo "valmin=$valmin"
-if [ "$valmin" != "" ]; then
-cat << EOF > $ini_filename
-V2_SLICE
- 0 $valmin 0 $valmax $slice_quantity_short
-
-EOF
-fi
-}
-
-
-#---------------------------------------------
-#                   select_slice
-#---------------------------------------------
-
-select_slice()
-{
-    have_bounds=
-    cat $slicefilemenu | awk -F"," '{ print $1" ",$2," ",$3," ",$4}'
-    read -p "Select slice index to generate an html file: " ans
-    if [[ "$ans" -ge 1 ]] && [[ "$ans" -le "$nslices" ]]; then
-      slice_index=$ans
-      slice_quantity=`cat $slicefilemenu | awk -v ind="$slice_index" -F"," '{ if($1 == ind){print $2} }'`
-      slice_quantity=`trim "$slice_quantity"`
-
-      slice_quantity_short=`grep -A 4 SLCF $smvfile | grep "$slice_quantity" -A 1 | head -2 | tail -1`
-      slice_quantity_short=`trim "$slice_quantity_short"`
-
-      slice_quantity_unit=`grep -A 4 SLCF $smvfile | grep "$slice_quantity" -A 2 | tail -1`
-      slice_quantity_unit=`trim "$slice_quantity_unit"`
-      restore_localstate $ans
-
-    else
-      echo slice index, $ans, is out of bounds
-    fi
+  if [ -e $GLOBALCONFIG ]; then
+    echo "#/bin/bash"                             >  $GLOBALCONFIG
+    echo "export SLICE2HTML_HTMLDIR=$HTMLDIRDIR"  >> $GLOBALCONFIG
+    echo "export SLICE2HTML_TIMEFRAME=$TIMEFRAME" >> $GLOBALCONFIG
+    echo "export SLICE2HTML_EMAIL=$EMAIL"         >> $GLOBALCONFIG
+  fi
 }
 
 #---------------------------------------------
@@ -156,51 +77,27 @@ select_slice()
 
 generate_html ()
 {
+script_index=$1
 while true; do
   echo ""
   echo "slice quantity: $slice_quantity "
-  if [ "$TIMEFRAME" == "-1" ]; then
-    echo " time frame(s):  all"
-  else
-    echo " time frame(s):  $TIMEFRAME"
-  fi
-if [ "$have_bounds" == "1" ]; then
-  echo "      min, max: $valmin $slice_quantity_unit, $valmax $slice_quantity_unit"
-else
-  echo "        bounds: default"
-fi
-
-  echo "      html dir: $HTMLDIR"
   echo "     smokeview: $SMOKEVIEW"
   echo "         email: $EMAIL"
   echo ""
-  echo "s - select slice"
-  echo "b - set bounds"
-  echo "h - set html directory"
-  echo "m - set email address"
-  echo "t - set time frame (-1 for all time frames)"
+  echo "h - define directory containing html page"
+  echo "m - select email address"
+  echo "t - select time frame (-1 for all time frames)"
   echo "1 - generate html page"
   echo "x - exit"
-
+  EXIT_SCRIPT=
   read -p "option: " ans
   if [ "$ans" == "h" ]; then
-    read -p "   enter html dir: " HTMLDIR
+    read -p "   enter html directory: " HTMLDIR
     CHECK_WRITE $HTMLDIR
     continue
   fi
-  if [ "$ans" == "b" ]; then
-    have_bounds=1
-    read -p "   set $slice_quantity_short min: " valmin
-    read -p "   set $slice_quantity_short max: " valmax
-    save_localstate $slice_index
-    continue
-  fi
-  if [ "$ans" == "s" ]; then
-    select_slice
-    continue
-  fi
   if [ "$ans" == "m" ]; then
-    read -p "   set email address: " EMAIL
+    read -p "   enter email address: " EMAIL
     continue
   fi
   if [ "$ans" == "t" ]; then
@@ -208,6 +105,7 @@ fi
     continue
   fi
   if [ "$ans" == "x" ]; then
+    EXIT_SCRIPT=1
     save_state
     exit
   fi
@@ -215,27 +113,20 @@ fi
     save_state
     if [ -e $scriptname ]; then
       echo "***creating html file"
-      scriptname=${SMVSCRIPTDIR}${input}_slice_${slice_index}.ssf
-      smv_inifilename=${SMVSCRIPTDIR}${input}_slice_${slice_index}.ini
-      writeini $smv_inifilename
+      scriptname=${SMVSCRIPTDIR}${input}_slice_${ans}.ssf
       rm -f $scriptname
-      GENERATE_SCRIPT $slice_index $scriptname $smv_inifilename
+      GENERATE_SCRIPT $script_index $scriptname
       echo dir=`pwd`
       echo "$SMOKEVIEW -htmlscript $scriptname -bindir $SMVBINDIR  $input"
             $SMOKEVIEW -htmlscript $scriptname -bindir $SMVBINDIR  $input
       if [ -e $htmlfile ]; then
         filesize=`ls -lk $htmlfile | awk '{print $5}'`
-        filesize=$((filesize/1000))
         echo "*** The html file, $htmlfile(${filesize}K), has been created."
-        if [[ "$EMAIL" != "" ]] && [[ -e $htmlfile ]]; then
-          echo "*** The html file, $htmlfile(${filesize}K), has been emailed to $EMAIL"
-          echo "" | mail -s "html page for $slice_quantity(${filesize}K)" -a $htmlfile $EMAIL
-        fi
       else
         echo "*** Creation of the html file $htmlfile failed."
       fi
     fi
-    continue
+    return
   fi
 done
 }
@@ -248,28 +139,16 @@ GENERATE_SCRIPT ()
 {
   ind=$1
   scriptname=$2
-  ini_filename=$3
   htmlbase=${input}_slice_${ind}
   htmlfile=$HTMLDIR/${htmlbase}.html
   cat << EOF > $scriptname
 RENDERHTMLDIR
   $HTMLDIR
 UNLOADALL
-LOADINIFILE  
- $ini_filename  
 LOADSLICE
 EOF
   cat $slicefilemenu | awk -v ind="$ind" -F"," '{ if($1 == ind){print $2"\n" $3 $4"\n"} }' >> $scriptname
-
   slice_quantity=`cat $slicefilemenu | awk -v ind="$slice_index" -F"," '{ if($1 == ind){print $2} }'`
-  slice_quantity=`trim "$slice_quantity"`
-
-  slice_quantity_short=`grep -A 4 SLCF $slicefilemenu | grep "$slice_quantity" -A 1 | head -2 | tail -1`
-  slice_quantity_short=`trim "$slice_quantity_short"`
-
-  slice_quantity_unit=`grep -A 4 SLCF $slicefilemenu | grep "$slice_quantity" -A 2 | tail -1`
-  slice_quantity_unit=`trim "$slice_quantity_unit"`
-
   cat << EOF >> $scriptname
 RENDERHTMLALL
   $htmlbase
@@ -288,10 +167,6 @@ SMVREPO=$ROOTDIR/smv
 BOTREPO=$ROOTDIR/bot
 cd $CURDIR
 SMOKEVIEW=$SMVREPO/Build/smokeview/intel_linux_64/smokeview_linux_64
-if [ ! -e $SMOKEVIEW ]; then
-  SMOKEVIEW=$SMVREPO/Build/smokeview/intel_linux_64/smokeview_linux_test_64
-fi
-
 SMVBINDIR=$BOTREPO/Bundle/smv/for_bundle
 
 CONFIGDIR=$HOME/.smokeview
@@ -333,7 +208,6 @@ case $OPTION  in
 esac
 done
 shift $(($OPTIND-1))
-
 restore_state
 
 
@@ -354,7 +228,6 @@ fi
 
 if [ ! -e $smvfile ]; then
   echo "***error: The .smv file, $smvfile, does not exist"
-  exit 1
 fi
 
 if [ ! -e $slicefilemenu ]; then
@@ -368,9 +241,27 @@ fi
 
 nslices=`cat $slicefilemenu | wc -l`
 if [ "$nslices" == "0" ]; then
-  echo "*** No slice files found in the smokeview file $smvfile"
+  echo "*** No slice files exist in the smokeview file $smvfile"
   exit
 fi
 
-select_slice
-generate_html
+while true; do
+  OUTPUT_SLICES
+  read -p "Select slice index to generate an html file: " ans
+  if [[ "$ans" -ge 1 ]] && [[ "$ans" -le "$nslices" ]]; then
+    slice_index=$ans
+    slice_quantity=`cat $slicefilemenu | awk -v ind="$slice_index" -F"," '{ if($1 == ind){print $2} }'`
+    generate_html $ans
+    if [ "$EXIT_SCRIPT" != "" ]; then
+      exit
+    fi
+    if [ "$EMAIL" != "" ]; then
+      if [ -e $htmlfile ]; then
+        echo "html file, $htmlfile, sent to $EMAIL"
+        echo "" | mail -s "html page for of $slice_quantity" -a $htmlfile $EMAIL
+      fi
+    fi
+  else
+    echo index $ans out of bounds
+  fi
+done
